requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
earthengine-api==0.1.355
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
oauth2client==4.1.3
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyDrive==1.3.1
pyparsing==3.0.9
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.16
click==8.1.7
loguru=0.7.2